+++ scripts/guacamole-spark-defaults.conf
@@ -1,4 +1,4 @@
 spark.serializer               org.apache.spark.serializer.KryoSerializer
-spark.kryo.registrator         org.hammerlab.guacamole.GuacamoleKryoRegistrator
+spark.kryo.registrator         org.hammerlab.guacamole.kryo.GuacamoleKryoRegistrator
 spark.kryoserializer.buffer.mb 4
 spark.kryo.referenceTracking   true
