+++ src/main/scala/org/hammerlab/guacamole/Common.scala
@@ -386,7 +386,7 @@ object Common extends Logging {
     }
 
     if (config.getOption("spark.kryo.registrator").isEmpty) {
-      config.set("spark.kryo.registrator", "org.hammerlab.guacamole.GuacamoleKryoRegistrator")
+      config.set("spark.kryo.registrator", "org.hammerlab.guacamole.kryo.GuacamoleKryoRegistrator")
     }
 
     if (config.getOption("spark.kryoserializer.buffer").isEmpty) {
