+++ src/test/scala/org/hammerlab/guacamole/util/GuacFunSuite.scala
@@ -8,7 +8,7 @@ trait GuacFunSuite extends SparkFunSuite with Matchers {
   override val properties: Map[String, String] =
     Map(
       "spark.serializer" -> "org.apache.spark.serializer.KryoSerializer",
-      "spark.kryo.registrator" -> "org.hammerlab.guacamole.GuacamoleKryoRegistrator",
+      "spark.kryo.registrator" -> "org.hammerlab.guacamole.kryo.GuacamoleKryoRegistrator",
       "spark.kryoserializer.buffer" -> "4",
       "spark.kryo.referenceTracking" -> "true"
     )
